Added initial coordinates, fixed balls gluing with walls

+++ src/main/java/ru/nsu/fit/akitov/billiards/utils/Point2D.java
@@ -1,4 +1,4 @@
 package ru.nsu.fit.akitov.billiards.utils;
 
-public record Point2D(int x, int y) {
+public record Point2D(float x, float y) {
 }
